requirements.txt
gradio
requests
langgraph
langchain
langchain-google-genai
pandas
tabulate
python-dotenv
google-generativeai